+++ datamgt/datasource/location_rating.csv
@@ -1,0 +1,151 @@
+﻿LOCATION,POSTCODE,RATING
+loc_1,DT,3
+loc_2,BH,1
+loc_3,SO,3
+loc_4,DT,4
+loc_5,BH,1
+loc_6,SO,3
+loc_7,DT,2
+loc_8,BH,2
+loc_9,SO,1
+loc_10,DT,1
+loc_11,BH,3
+loc_12,SO,3
+loc_13,DT,3
+loc_14,BH,1
+loc_15,SO,3
+loc_16,DT,2
+loc_17,BH,3
+loc_18,SO,2
+loc_19,DT,2
+loc_20,BH,2
+loc_21,SO,1
+loc_22,DT,1
+loc_23,BH,4
+loc_24,SO,4
+loc_25,DT,3
+loc_26,BH,1
+loc_27,SO,1
+loc_28,DT,2
+loc_29,BH,3
+loc_30,SO,3
+loc_31,DT,4
+loc_32,BH,3
+loc_33,SO,4
+loc_34,DT,2
+loc_35,BH,2
+loc_36,SO,4
+loc_37,DT,2
+loc_38,BH,1
+loc_39,SO,4
+loc_40,DT,4
+loc_41,BH,4
+loc_42,SO,2
+loc_43,DT,3
+loc_44,BH,4
+loc_45,SO,1
+loc_46,DT,3
+loc_47,BH,1
+loc_48,SO,4
+loc_49,DT,2
+loc_50,BH,4
+loc_51,SO,4
+loc_52,DT,3
+loc_53,BH,4
+loc_54,SO,3
+loc_55,DT,1
+loc_56,BH,4
+loc_57,SO,2
+loc_58,DT,4
+loc_59,BH,4
+loc_60,SO,2
+loc_61,DT,2
+loc_62,BH,3
+loc_63,SO,3
+loc_64,DT,2
+loc_65,BH,1
+loc_66,SO,2
+loc_67,DT,3
+loc_68,BH,3
+loc_69,SO,4
+loc_70,DT,4
+loc_71,BH,3
+loc_72,SO,2
+loc_73,DT,4
+loc_74,BH,1
+loc_75,SO,1
+loc_76,DT,3
+loc_77,BH,4
+loc_78,SO,1
+loc_79,DT,1
+loc_80,BH,3
+loc_81,SO,1
+loc_82,DT,2
+loc_83,BH,2
+loc_84,SO,2
+loc_85,DT,4
+loc_86,BH,1
+loc_87,SO,2
+loc_88,DT,2
+loc_89,BH,4
+loc_90,SO,1
+loc_91,DT,4
+loc_92,BH,3
+loc_93,SO,4
+loc_94,DT,3
+loc_95,BH,3
+loc_96,SO,3
+loc_97,DT,3
+loc_98,BH,4
+loc_99,SO,2
+loc_100,DT,3
+loc_101,BH,4
+loc_102,SO,4
+loc_103,DT,3
+loc_104,BH,4
+loc_105,SO,2
+loc_106,DT,1
+loc_107,BH,1
+loc_108,SO,2
+loc_109,DT,1
+loc_110,BH,4
+loc_111,SO,4
+loc_112,DT,3
+loc_113,BH,2
+loc_114,SO,4
+loc_115,DT,4
+loc_116,BH,4
+loc_117,SO,1
+loc_118,DT,4
+loc_119,BH,2
+loc_120,SO,2
+loc_121,DT,2
+loc_122,BH,1
+loc_123,SO,4
+loc_124,DT,3
+loc_125,BH,4
+loc_126,SO,3
+loc_127,DT,1
+loc_128,BH,1
+loc_129,SO,3
+loc_130,DT,1
+loc_131,BH,3
+loc_132,SO,2
+loc_133,DT,1
+loc_134,BH,2
+loc_135,SO,4
+loc_136,DT,1
+loc_137,BH,1
+loc_138,SO,1
+loc_139,DT,4
+loc_140,BH,2
+loc_141,SO,3
+loc_142,DT,3
+loc_143,BH,3
+loc_144,SO,4
+loc_145,DT,2
+loc_146,BH,2
+loc_147,SO,2
+loc_148,DT,1
+loc_149,BH,4
+loc_150,SO,3
