+++ datamgt/drafts/JSONEATER.PY
@@ -1,0 +1,6 @@
+import json
+
+
+
+with open("CQC_data_file_multiples.json", "r") as read_file:
+    data = json.load(read_file)
