--- datamgt/helpers/drafts/JSONEATER.PY
@@ -1,6 +1,0 @@
-import json
-
-
-
-with open("CQC_data_file_multiples.json", "r") as read_file:
-    data = json.load(read_file)
